MTD-24- Updated Tower Ranges
Showing tower ranges now handled in UI Stage
