+++ core/src/com/eric/mtd/game/ui/presenter/SupportPresenter.java
@@ -137,22 +137,6 @@ public class SupportPresenter implements IGameUIStateObserver {
 			
 		}
 	}
-	
-	/**
-	 * Show/Hide tower ranges for all towers
-	 * 
-	 * @param showRanges
-	 */
-	public void showTowerRanges(boolean showRanges) {
-		if (showRanges){
-			System.out.println("Showing ranges");
-			for (Actor tower : actorGroups.getTowerGroup().getChildren()) {
-				if (tower instanceof Tower) {
-					((CombatActor) tower).setShowRange(showRanges);
-				}
-			}
-		}
-	}
 
 	/**
 	 * Determines if the tower can be purchased.
@@ -188,19 +172,15 @@ public class SupportPresenter implements IGameUIStateObserver {
 		switch (state) {
 		case SUPPORT:
 			view.supportState();
-			showTowerRanges(false);
 			break;
 		case PLACING_SUPPORT:
 			view.placingSupportState();
-			showTowerRanges(true);
 			break;
 		case PLACING_AIRSTRIKE:
 			view.placingSupportState();
-			showTowerRanges(true);
 			break;
 		default:
 			view.standByState();
-			showTowerRanges(false);
 			break;
 		}
